Fix Windows: 64-bit source IDs and POSIX path serialization
Two platform-ABI bugs surfaced by the Windows CI cell (both pre-existing,
not numpy-2 regressions):

Source IDs: Gaia source ids are ~60-bit, but were stored with the
platform-dependent numpy.uint / numpy.dtype(int), which is 32-bit on
Windows (LLP64). The ids truncated and collided, so TFA's "which template
is this same star" match hit multiple entries and int(numpy.nonzero(...))
got a multi-element array. Use numpy.uint64 for the source-id arrays
(_init_magfit_sources, lc_detrending ID field) and make the scalar
extraction numpy-2-clean.

Paths: the error sidecar/detail serialized paths with str(Path), which
emits backslashes on Windows. Use Path.as_posix() in to_detail_dict,
sanitize_for_json, and the artifact-FK resolution so serialized and
compared paths are platform-stable.

+++ autowisp/magnitude_fitting/util.py
@@ -27,7 +27,9 @@ def _init_magfit_sources(source_data):
 
     dtype = []
     if "source_id" not in source_data:
-        dtype.append(("source_id", numpy.uint, 3))
+        # uint64, not the platform-dependent numpy.uint (32-bit on Windows):
+        # the middle component holds the full Gaia source id (~60 bits).
+        dtype.append(("source_id", numpy.uint64, 3))
 
     num_phot = 0
     found_magfit_iterations = set()
